+++ sklearn/ensemble/__init__.py
@@ -4,13 +4,13 @@ classification and regression.
 """
 
 from .base import BaseEnsemble
-from .bagging import BaggingClassifier
-from .bagging import BaggingRegressor
 from .forest import RandomForestClassifier
 from .forest import RandomForestRegressor
 from .forest import RandomTreesEmbedding
 from .forest import ExtraTreesClassifier
 from .forest import ExtraTreesRegressor
+from .random_patches import RandomPatchesClassifier
+from .random_patches import RandomPatchesRegressor
 from .weight_boosting import AdaBoostClassifier
 from .weight_boosting import AdaBoostRegressor
 from .gradient_boosting import GradientBoostingClassifier
@@ -21,10 +21,11 @@ from . import weight_boosting
 from . import gradient_boosting
 from . import partial_dependence
 
-__all__ = ["BaseEnsemble", "BaggingClassifier", "BaggingRegressor",
+__all__ = ["BaseEnsemble",
            "RandomForestClassifier", "RandomForestRegressor",
            "RandomTreesEmbedding", "ExtraTreesClassifier",
-           "ExtraTreesRegressor", "GradientBoostingClassifier",
+           "ExtraTreesRegressor", "RandomPatchesClassifier",
+           "RandomPatchesRegressor", "GradientBoostingClassifier",
            "GradientBoostingRegressor", "AdaBoostClassifier",
            "AdaBoostRegressor", "forest", "gradient_boosting",
            "partial_dependence", "weight_boosting"]
