+++ sklearn/ensemble/random_patches.py
