+++ sklearn/ensemble/tests/test_random_patches.py
